Added fixes to conftest folder

+++ tests/conftest.py
@@ -2,6 +2,7 @@ import pytest
 import simvue
 import uuid
 import subprocess
+import time
 
 @pytest.fixture(scope='session', autouse=True)
 def folder_setup():
@@ -10,7 +11,10 @@ def folder_setup():
     yield folder
     # Will be executed after the last test
     client = simvue.Client()
-    client.delete_folder(folder, remove_runs=True)
+    if client.get_folder(folder):
+        # Avoid trying to delete folder while one of the runs is still closing
+        time.sleep(1)
+        client.delete_folder(folder, remove_runs=True)
     
     
 @pytest.fixture()
